backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.1
pydantic==2.9.2
pydantic-settings==2.5.2
supabase==2.5.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
python-dateutil==2.9.0.post0
pytz==2024.1
slowapi==0.1.9
pytest==8.2.0
pytest-cov==5.0.0
black==24.4.2
flake8==7.0.0
